refactor: relocate browser object to ports/_browser/_raw.py
- Create webcompy/ports/_browser/_raw.py with browser object definition
- Create webcompy/ports/_browser/_raw.pyi with full BrowserModule Protocol type stubs
- Update 7 port implementation imports to _raw path
- Replace _browser/_modules.py with re-export stub (Router compat, phase 6)
- Remove public browser export from webcompy/__init__.py
- Empty _browser/__init__.py
- Update pyproject.toml stubPath to webcompy/ports
- Delete _browser/_modules.pyi (superseded by ports/_browser/_raw.pyi)
- Add DOMPort add/remove_document_event_listener methods
- Fix docs_app components (navigation → DI, hljs → _raw)
- Update feat-remove-browser proposal and tasks
