+++ docs_app/components/demo_display.py
@@ -1,10 +1,10 @@
 from typing import TypedDict
 
-from webcompy import browser
 from webcompy.aio import AsyncWrapper
 from webcompy.ajax import HttpClient
 from webcompy.components import ComponentContext, define_component
 from webcompy.elements import DomNodeRef, html
+from webcompy.ports._browser._raw import browser
 from webcompy.signal import Signal
 from webcompy.utils import strip_multiline_text
