+++ webcompy/_browser/__init__.py
@@ -1,5 +1,1 @@
-from webcompy._browser._modules import browser
 
-__all__ = [
-    "browser",
-]
