+++ webcompy/_browser/_modules.py
@@ -1,6 +1,1 @@
-from webcompy.utils._environment import ENVIRONMENT as _ENVIRONMENT
-
-if _ENVIRONMENT == "pyscript":
-    from webcompy._browser._pyscript import browser  # type: ignore
-else:
-    browser = None
+from webcompy.ports._browser._raw import browser  # noqa: F401
