+++ webcompy/ports/_browser/_raw.py
@@ -1,0 +1,6 @@
+from webcompy.utils._environment import ENVIRONMENT as _ENVIRONMENT
+
+if _ENVIRONMENT == "pyscript":
+    from webcompy._browser._pyscript import browser  # type: ignore
+else:
+    browser = None
