+++ webcompy/ports/_dom.py
@@ -84,3 +84,7 @@ class DOMPort(ABC):
     def set_title(self, title: str) -> None: ...
     @abstractmethod
     def schedule_macro_task(self, callback: Any) -> None: ...
+    @abstractmethod
+    def add_document_event_listener(self, event_type: str, handler: Any) -> None: ...
+    @abstractmethod
+    def remove_document_event_listener(self, event_type: str, handler: Any) -> None: ...
